- move evolution logic and legendary filter to its own services - fix evolution method filters for a few Pokémon
