+++ src/components/SelectEvolveModal.tsx
@@ -1,7 +1,4 @@
 import React, { useEffect, useId, useMemo, useRef, useState } from "react";
-import { TFunction } from "i18next";
-import { EVOLUTIONS_DE, EVOLUTIONS_EN } from "@/src/data/pokemon-evolutions";
-import { POKEMON_ID_TO_GENERATION } from "@/src/data/pokemon-map";
 import {
   getOfficialArtworkUrlById,
   getSpriteUrlById,
@@ -12,14 +9,12 @@ import {
 } from "@/src/services/pokemonSearch";
 import type { PokemonLink } from "@/types";
 import { useTranslation } from "react-i18next";
-import {
-  normalizeLanguage,
-  type SupportedLanguage,
-} from "@/src/utils/language";
+import { normalizeLanguage } from "@/src/utils/language";
 import { focusRingClasses } from "@/src/styles/focusRing";
 import { useFocusTrap } from "@/src/hooks/useFocusTrap";
 import { getPokemonTypeSlugsById } from "@/src/services/pokemonTypes";
 import TypeBadge from "@/src/components/TypeBadge";
+import { getFilteredEvolutionEntriesForPokemon } from "@/src/services/evolutionMethodFilter.ts";
 
 interface SelectEvolveModalProps {
   isOpen: boolean;
@@ -39,14 +34,6 @@ interface EvoInfo {
   methods: string[];
   artworkUrl?: string | null;
 }
-
-const EVOLUTION_TABLES: Record<
-  SupportedLanguage,
-  Record<number, { id: number; methods: string[] }[]>
-> = {
-  de: EVOLUTIONS_DE,
-  en: EVOLUTIONS_EN,
-};
 
 function mergeLocationMethods(
   methods: string[] | undefined,
@@ -70,178 +57,6 @@ function mergeLocationMethods(
   if (locations.length === 1)
     return [...withoutPlainLevelUp, `${locationPrefix}${locations[0]}`];
   return [...withoutPlainLevelUp, `${locationPrefix}${locations.join(", ")}`];
-}
-
-interface MethodGenerationRule {
-  methods: string[];
-  minGeneration?: number;
-  maxGeneration?: number;
-}
-
-// id = pokemon which evolution results in
-const METHOD_GENERATION_RULES: Record<number, MethodGenerationRule[]> = {
-  28: [
-    {
-      methods: ["Item: Eisstein", "Item: Ice Stone"],
-      minGeneration: 7,
-    },
-  ],
-  38: [
-    {
-      methods: ["Item: Eisstein", "Item: Ice Stone"],
-      minGeneration: 7,
-    },
-  ],
-  20: [
-    {
-      methods: [
-        "Level-Up - Level 20, Tageszeit: Nacht",
-        "Level-Up - Level 20, Time of day: Night",
-      ],
-      minGeneration: 7,
-    },
-  ],
-  53: [
-    {
-      methods: ["Level-Up - Freundschaft ≥ 160", "Level-Up - Friendship ≥ 160"],
-      minGeneration: 7,
-    },
-  ],
-  700: [
-    {
-      methods: [
-        "Level-Up - Freundschaft ≥ 160, Kennt Attacke vom Typ Fee",
-        "Level-Up - Friendship ≥ 160, Knows move of type Fairy",
-      ],
-      minGeneration: 8,
-    },
-  ],
-  80: [
-    {
-      methods: ["Item: Galarnuss-Reif", "Item: Galarica Cuff"],
-      minGeneration: 7,
-    },
-  ],
-  199: [
-    {
-      methods: ["Item: Galarnuss-Kranz", "Item: Galarica Wreath"],
-      minGeneration: 7,
-    },
-  ],
-  105: [
-    {
-      methods: [
-        "Level-Up - Level 28, Tageszeit: Nacht",
-        "Level-Up - Level 28, Time of day: Night",
-      ],
-      minGeneration: 7,
-    },
-  ],
-};
-
-const LOCATION_METHOD_VERSION_RULES: Record<string, string[]> = {
-  "Level-Up - Ort: Kraterberg": ["gen4_dp", "gen4_pt"],
-  "Level-Up - Location: Mt. Coronet": ["gen4_dp", "gen4_pt"],
-  "Level-Up - Ort: Elektrolithhöhle": ["gen5_sw", "gen5_s2w2"],
-  "Level-Up - Location: Chargestone Cave": ["gen5_sw", "gen5_s2w2"],
-  "Level-Up - Ort: Route 13": ["gen6_xy"],
-  "Level-Up - Location: Route 13": ["gen6_xy"],
-  "Level-Up - Ort: Ewigenwald": ["gen5_sw", "gen5_s2w2"],
-  "Level-Up - Location: Pinwheel Forest": ["gen5_sw", "gen5_s2w2"],
-  "Level-Up - Ort: Ewigwald": ["gen4_dp", "gen4_pt"],
-  "Level-Up - Location: Eterna Forest": ["gen4_dp", "gen4_pt"],
-  "Level-Up - Ort: Route 20": ["gen6_xy"],
-  "Level-Up - Location: Route 20": ["gen6_xy"],
-  "Level-Up - Ort: Route 217": ["gen4_dp", "gen4_pt"],
-  "Level-Up - Location: Route 217": ["gen4_dp", "gen4_pt"],
-  "Level-Up - Ort: Wendelberg": ["gen5_sw", "gen5_s2w2"],
-  "Level-Up - Location: Twist Mountain": ["gen5_sw", "gen5_s2w2"],
-  "Level-Up - Ort: Frosthöhle": ["gen6_xy"],
-  "Level-Up - Location: Frost Cavern": ["gen6_xy"],
-};
-
-function methodAllowedForVersion(method: string, gameVersionId?: string) {
-  if (!gameVersionId) return true;
-  const allowedVersions = LOCATION_METHOD_VERSION_RULES[method];
-  if (!allowedVersions) return true;
-  return allowedVersions.includes(gameVersionId);
-}
-
-function filterMethodsForConstraints(
-  pokemonId: number,
-  methods: string[] | undefined,
-  maxGeneration?: number | null,
-  gameVersionId?: string,
-) {
-  if (!methods || methods.length === 0) return methods || [];
-  let filtered = methods;
-  if (typeof maxGeneration === "number") {
-    const rules = METHOD_GENERATION_RULES[pokemonId];
-    if (rules && rules.length) {
-      filtered = filtered.filter((method) => {
-        const rule = rules.find((r) => r.methods?.includes(method));
-        if (!rule) return true;
-        if (
-          typeof rule.minGeneration === "number" &&
-          maxGeneration < rule.minGeneration
-        )
-          return false;
-        if (
-          typeof rule.maxGeneration === "number" &&
-          maxGeneration > rule.maxGeneration
-        )
-          return false;
-        return true;
-      });
-    }
-  }
-  if (gameVersionId) {
-    filtered = filtered.filter((method) =>
-      methodAllowedForVersion(method, gameVersionId),
-    );
-  }
-  return filtered;
-}
-
-function getFilteredEvolutionEntriesForPokemon(
-  pokemonId: number,
-  language: SupportedLanguage,
-  t: TFunction,
-  maxGeneration?: number | null,
-  gameVersionId?: string,
-) {
-  const evoTable = EVOLUTION_TABLES[language] || EVOLUTION_TABLES.de;
-  const evoEntries = evoTable[pokemonId] || [];
-  const generationLimit =
-    typeof maxGeneration === "number" ? maxGeneration : null;
-  const filteredEntries =
-    generationLimit === null
-      ? evoEntries
-      : evoEntries.filter((entry) => {
-          const gen = POKEMON_ID_TO_GENERATION[entry.id];
-          if (typeof gen !== "number") return true;
-          return gen <= generationLimit;
-        });
-  if (filteredEntries.length === 0) return [];
-  const defaultUnknown = t("tracker.evolveModal.methodUnknown");
-  return filteredEntries
-    .map((entry) => {
-      const baseMethods =
-        entry.methods && entry.methods.length
-          ? entry.methods
-          : [defaultUnknown];
-      const methodsForConstraints = filterMethodsForConstraints(
-        entry.id,
-        baseMethods,
-        maxGeneration,
-        gameVersionId,
-      );
-      return {
-        id: entry.id,
-        methods: methodsForConstraints,
-      };
-    })
-    .filter((entry) => entry.methods.length > 0);
 }
 
 const SelectEvolveModal: React.FC<SelectEvolveModalProps> = ({
